Changed size of size_t to uint32_t

+++ include/common/types.h
@@ -16,7 +16,7 @@ typedef unsigned int uint32_t;
 typedef long long int int64_t;
 typedef unsigned long long int uint64_t;
 
-typedef unsigned long int size_t;
+typedef uint32_t size_t;
 
 } // namespace common
 } // namespace morphios
